refactor: 보드의 생성 책임 팩토리 객체로 이전
